+++ changelog.md
@@ -1,6 +1,7 @@
 - **2nd August, 2023 - Eric McWinNEr 😴**
   - Made bugfixes to the responsive layout
   - Fixed bug with the export csv feature on the table
+  - Fixed bug with the generated csv name on the table for the csv export
 - **1st August, 2023 - Eric McWinNEr 😴**
   - Made minor update to the pin input, tag input and tag dropdown
   - Made the right icon of the select field customizable